We can read atoms and stuff and do put class!!!11

+++ PDBReader/Coordinates/Vector3.cs
@@ -3,11 +3,11 @@
     internal class Vector3
     {
 
-        float x;
-        float y;
-        float z;
+        double x;
+        double y;
+        double z;
 
-        public Vector3(float x, float y, float z)
+        public Vector3(double x, double y, double z)
         {
             this.x = x;
             this.y = y;
